// apps/ui/admin/src/models/dataStore.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * wanaku-barn-backend API
 * OpenAPI spec version: 0.3.0-SNAPSHOT
 */
import type { DataStoreLabels } from "./dataStoreLabels";

export interface DataStore {
  labels?: DataStoreLabels;
  id?: string;
  name?: string;
  data?: string;
}